LPS-92231 Initialize experiences only with standard content pages
